Move APIHandler fns into folder
